SignalCluster: fixed broken mobile icons on some MTK devices
All fields were being stored on the first run and that was breaking
mobile icons coloring and even causing data activity and mobile type not
to show up on some MTK devices. So, not all field can be stored because
they are dynamic and can change on each method call.