QuietHours: implemented Wear mode
When this mode is active, all notification sounds and device-wide
vibrations are muted on the device (including apps having ignore rules)
while vibrations on a connected watch are maintained. Useful during work
hours when device should remain silent while watch is able to notify of
new events using its own vibrations.
Mode can be toggled using existing QuietHours tile or GravityBox Action
(shortcut) and is available only when Android Wear app is present on the
device which is a good indication that the device is actually paired
with a watch.
Closes #1639
# Conflicts:
#	src/com/ceco/lollipop/gravitybox/StatusbarQuietHoursView.java
#	src/com/ceco/lollipop/gravitybox/ledcontrol/QuietHoursActivity.java
#	src/com/ceco/lollipop/gravitybox/quicksettings/QuietHoursTile.java
#	src/com/ceco/lollipop/gravitybox/shortcuts/QuietHoursShortcut.java

# Conflicts:
#	src/com/ceco/kitkat/gravitybox/StatusbarQuietHoursView.java
#	src/com/ceco/kitkat/gravitybox/ledcontrol/QuietHoursActivity.java
#	src/com/ceco/kitkat/gravitybox/shortcuts/QuietHoursShortcut.java
#	src/com/ceco/lollipop/gravitybox/quicksettings/QuietHoursTile.java