NFC: improved reliability of NFC related features
Fixes NFC tile as well as NFC related GravityBox actions not working
sometimes.
We use deprecated getNfcAdapter() with null context that's available for
backwards compatibility since we do not need to use NFC API that require
application context, anyway. We only need to check for NFC state and
call enable/disable on adapter.